requirements.txt
anthropic>=0.40.0
watchdog>=4.0.0
PyMuPDF>=1.24.0
pdf2image>=1.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
rapidfuzz>=3.0.0
pydantic>=2.0.0
PyYAML>=6.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
click>=8.1.0
schedule>=1.2.0
python-magic>=0.4.27
flask>=3.0.0
